Create field group with text field input
